Two fixes w.r.t. enum strings builder: - They are now always linked to binary files first so that they are available when rrlib_rtti data types are initialized. - GetEnumString() C++ function can now handle enum classes.

+++ enum_strings_builder/enum_strings.h
@@ -94,7 +94,7 @@ const char* GetEnumString(ENUM value)
   const std::vector<const char*>* strings = GetEnumStrings<ENUM>();
   if (strings != NULL)
   {
-    assert(value < strings->size());
+    assert(static_cast<size_t>(value) < strings->size());
     return (*strings)[static_cast<size_t>(value)];
   }
   return NULL;
